+++ terra/compute/base.py
@@ -1,10 +1,15 @@
 import os
 import json
+import time
+import atexit
+from logging import StreamHandler
+from logging.handlers import SocketHandler
+import threading
 
 from terra import settings
 import terra.compute.utils
 from terra.executor import Executor
-from terra.logger import getLogger
+from terra.logger import getLogger, LogRecordSocketReceiver
 logger = getLogger(__name__)
 
 
@@ -186,5 +191,89 @@ class BaseCompute:
 
     return service_info.volumes
 
+  @staticmethod
+  def configure_logger(sender, **kwargs):
+    if settings.terra.zone == 'controller':
+      print("SGR - setting up controller logging")
+
+      # Setup log file for use in configure
+      sender._log_file = os.path.join(settings.processing_dir,
+                                      terra.logger._logs.default_log_prefix)
+      os.makedirs(settings.processing_dir, exist_ok=True)
+      sender._log_file = open(sender._log_file, 'a')
+      sender.main_log_handler = StreamHandler(stream=sender._log_file)
+      sender.root_logger.addHandler(sender.main_log_handler)
+
+
+      # setup the TCP socket listener
+      sender.tcp_logging_server = LogRecordSocketReceiver(
+          settings.logging.server.hostname, settings.logging.server.port)
+      print('SGR - About to start TCP server...')
+
+      listener_thread = threading.Thread(
+          target=sender.tcp_logging_server.serve_until_stopped)
+      listener_thread.setDaemon(True)
+      listener_thread.start()
+
+      # Wait up to a second, to make sure the thread started
+      for _ in range(1000):
+        if sender.tcp_logging_server.ready:
+          break
+        time.sleep(0.001)
+
+      # Auto cleanup
+      @atexit.register
+      def cleanup_thread():
+        print("SGR - Sending cease and desist to LogRecordSocketReceiver")
+        sender.tcp_logging_server.abort = 1
+        listener_thread.join(timeout=5)
+        if listener_thread.is_alive():
+          print("SGR - LogRecordSocketReceiver thread did not die")
+        print("SGR - LogRecordSocketReceiver died!")
+    elif settings.terra.zone == 'runner':
+      print(f"SGR - setting up BaseCompute:runner logging")
+
+      sender.main_log_handler = SocketHandler(
+          settings.logging.server.hostname, settings.logging.server.port)
+
+      # Now in configure_logger, you are able to access settings and determine
+      # whether there should be a stderr handler or not. If you don't so this,
+      # both the master controller and service runner will output the same log
+      # messages, duplicating output on stderr.
+      sender.root_logger.removeHandler(sender.stderr_handler)
+      # Some executors may need to do this too.
+
+  @staticmethod
+  def reconfigure_logger(sender, **kwargs):
+    # sender is logger in this case
+    #
+    # The default logging handler is a StreamHandler. This will reconfigure its
+    # output stream
+
+    if settings.terra.zone == 'controller':
+      print("SGR - BaseComputer:controller reconfigure logging")
+      log_file = os.path.join(settings.processing_dir,
+                              terra.logger._logs.default_log_prefix)
+
+      # if not os.path.samefile(log_file, sender._log_file.name):
+      if log_file != sender._log_file.name:
+        os.makedirs(settings.processing_dir, exist_ok=True)
+        sender._log_file.close()
+        sender._log_file = open(log_file, 'a')
+    elif settings.terra.zone == 'runner':
+      print("SGR - BaseComputer:runner reconfigure logging")
+      # Only if it's changed
+      if settings.logging.server.hostname != sender.main_log_handler.host or \
+         settings.logging.server.port != sender.main_log_handler.port:
+        # Reconnect Socket Handler
+        sender.main_log_handler.close()
+        try:
+          sender.root_logger.removeHandler(sender.main_log_handler)
+        except ValueError:
+          print(f"This shouldn't be happening...")
+          pass
+
+        sender.main_log_handler = SocketHandler(
+            settings.logging.server.hostname, settings.logging.server.port)
 
 services = {}
