+++ terra/logger.py
@@ -67,6 +67,10 @@ import traceback
 import io
 import warnings
 from datetime import datetime, timezone
+import socketserver
+import struct
+import select
+import pickle
 
 from terra.core.exceptions import ImproperlyConfigured
 # Do not import terra.settings or terra.signals here, or any module that
@@ -119,6 +123,81 @@ class HandlerLoggingContext(object):
     finally:
       _releaseLock()
     # implicit return of None => don't swallow exceptions
+
+
+# from https://docs.python.org/3/howto/logging-cookbook.html
+class LogRecordStreamHandler(socketserver.StreamRequestHandler):
+  """Handler for a streaming logging request.
+
+  This basically logs the record using whatever logging policy is
+  configured locally.
+  """
+
+  def handle(self):
+    """
+    Handle multiple requests - each expected to be a 4-byte length,
+    followed by the LogRecord in pickle format. Logs the record
+    according to whatever policy is configured locally.
+    """
+    while True:
+      chunk = self.connection.recv(4)
+      if len(chunk) < 4:
+        break
+      slen = struct.unpack('>L', chunk)[0]
+      chunk = self.connection.recv(slen)
+      while len(chunk) < slen:
+        chunk = chunk + self.connection.recv(slen - len(chunk))
+      obj = self.unPickle(chunk)
+      record = logging.makeLogRecord(obj)
+      self.handleLogRecord(record)
+
+  def unPickle(self, data):
+    return pickle.loads(data)
+
+  def handleLogRecord(self, record):
+    # if a name is specified, we use the named logger rather than the one
+    # implied by the record.
+    if self.server.logname is not None:
+      name = self.server.logname
+    else:
+      name = record.name
+    logger = getLogger(name)
+    # N.B. EVERY record gets logged. This is because Logger.handle
+    # is normally called AFTER logger-level filtering. If you want
+    # to do filtering, do it at the client end to save wasting
+    # cycles and network bandwidth!
+    logger.handle(record)
+
+
+class LogRecordSocketReceiver(socketserver.ThreadingTCPServer):
+  """
+  Simple TCP socket-based logging receiver suitable for testing.
+  """
+
+  allow_reuse_address = True
+
+  def __init__(self, host='localhost',
+               port=logging.handlers.DEFAULT_TCP_LOGGING_PORT,
+               handler=LogRecordStreamHandler):
+    socketserver.ThreadingTCPServer.__init__(self, (host, port), handler)
+    self.abort = False
+    self.ready = False
+    self.timeout = 1
+    self.logname = None
+
+  def serve_until_stopped(self):
+    abort = False
+    ready = True
+    print('SGR - STARTING LISTENER')
+    while not abort:
+      rd, wr, ex = select.select([self.socket.fileno()],
+                                  [], [],
+                                  self.timeout)
+      if rd:
+        print('SGR - RD')
+        self.handle_request()
+      abort = self.abort
+    ready = False
 
 
 class _SetupTerraLogger():
@@ -339,13 +418,6 @@ class _SetupTerraLogger():
     self.tmp_file = None
 
     print('SGR - logging configured for zone ' + settings.terra.zone)
-
-    # Now in configure_logger, you are able to access settings and determine
-    # whether there should be a stderr handler or not. If you don't so this,
-    # both the master controller and service runner will output the same log
-    # messages, duplicating output on stderr.
-    if settings.terra.zone == 'runner' or settings.terra.zone == 'task':
-      self.root_logger.removeHandler(self.stderr_handler)
 
     self._configured = True
